# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Roomrent API",
        "version": "1.0"
    },
    "host": "roomrent.dev",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/post": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "Post according to query string",
                "description": "gets post for loggedin user",
                "operationId": "getPost",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "gets the data skipping no of offset posts",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "user",
                        "in": "query",
                        "description": "gets the post of user",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "offer_or_ask",
                        "in": "query",
                        "description": "gets the offer or ask of user",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid inputs"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/post/create": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "creates post",
                "description": "creates post for loggedin user",
                "operationId": "setPost",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "title of post",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "post_description",
                        "in": "formData",
                        "description": "description of the post",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "location",
                        "in": "formData",
                        "description": "Address",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "latitude",
                        "in": "formData",
                        "description": "latitude of the location",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "longitude",
                        "in": "formData",
                        "description": "longitude of the place",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "price of the post",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "no_of_rooms",
                        "in": "formData",
                        "description": "number of rooms",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "offer_or_ask",
                        "in": "formData",
                        "description": "1 for offer 2 for ask",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "file[0]",
                        "in": "formData",
                        "description": "image file",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "file[1]",
                        "in": "formData",
                        "description": "image file",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid inputs"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/postbylocation": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "post near the given location",
                "description": "gets post around a certain distance",
                "operationId": "getPostByLocation",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "formData",
                        "description": "latitude of a place",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "longitude",
                        "in": "formData",
                        "description": "longitude of a place",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "offset",
                        "in": "formData",
                        "description": "gets post after offset value",
                        "type": "number"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid inputs"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}